+++ point/point.h
@@ -10,12 +10,12 @@
 #define point_h
 
 typedef struct point_t {
-    int x;
-    int y;
+    uint8_t x;
+    uint8_t y;
 } point_t;
 
 typedef struct point_namespace {
-    point_t* (*const construct)(int x, int y);
+    point_t* (*const construct)(uint8_t x, uint8_t y);
     int (*const destruct)(point_t* point);
     int (*const distance)(point_t* p1, point_t* p2);
     void (*const print)(point_t* p);
